backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
beautifulsoup4
playwright
scikit-learn
pandas
python-dotenv
apscheduler
